requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# API Clients
aiohttp==3.9.1
httpx==0.25.1
librouteros==3.3.0

# Async & Performance
asyncio==3.4.3
aiosignal==1.3.1
aiofiles==23.2.1

# Caching & Performance
redis==5.0.1
aioredis==2.0.1

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0

# Development
python-dotenv==1.0.0